requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic<2.0.0
requests
python-dotenv==0.19.0
streamlit<1.30.0
plotly==5.6.0
pandas<2.0.0
protobuf==3.20.3
# numpy is handled in code due to compatibility issues
python-multipart==0.0.5
networkx
psutil
pytest 